test: split response history coverage
